Remove BlockingUtils#asIO & monix-bio optional dependency
Recently, we've ran into a problem, where some usages of `BlockingUtils` which involved Java code would not respect the optionality of monix-bio on the classpath. Since there's only a single usage for monix-bio in the commons library and the maintenance status of the library is questionable, I'd like to propose removing the optional dependency. Removed method can be easily restored in the projects if at all needed.